src/backend/user/requirements.dev.txt
fastapi
sqlalchemy
pydantic
python-dotenv
alembic
uvicorn
asyncpg
psycopg2-binary
passlib
python-jose
minio
python-multipart
ruff
pydantic[email]
email-validator
confluent-kafka